# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exam-rotations:
    get:
      operationId: listExamRotations
      tags: [rotations]
      responses:
        "200":
          description: Exam rotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamRotation"
    post:
      operationId: createExamRotation
      tags: [rotations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRotationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamRotation"
  /exam-rotations/{id}:
    delete:
      operationId: deleteExamRotation
      tags: [rotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    patch:
      operationId: updateExamRotation
      tags: [rotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRotationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamRotation"
  /exam-rotations/{id}/assignments:
    get:
      operationId: listExamAssignments
      tags: [rotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamTeacherAssignment"
    post:
      operationId: saveExamAssignments
      tags: [rotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignments:
                  type: array
                  items:
                    $ref: "#/components/schemas/ExamAssignmentInput"
              required: [assignments]
      responses:
        "201":
          description: Saved
...
  /students/{id}/goals:
    get:
      operationId: listStudentGoals
      tags: [students]
      summary: List student goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentGoal"
    post:
      operationId: createStudentGoal
      tags: [students]
      summary: Create a student goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentGoalBody"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentGoal"
  /students/{id}/goals/{goalId}:
    patch:
      operationId: updateStudentGoal
      tags: [students]
      summary: Update a student goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentGoalBody"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentGoal"
    delete:
      operationId: deleteStudentGoal
      tags: [students]
      summary: Delete a student goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Goal deleted
...
components:
...
  schemas:
...
    StudentGoal:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        title:
          type: string
        targetDate:
          type: string
        notes:
          type: ["string", "null"]
        motivationalMessage:
          type: ["string", "null"]
        createdById:
          type: integer
        isCompleted:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
      required: [id, studentId, title, createdById, isCompleted, createdAt]
...
    CreateStudentGoalBody:
      type: object
      properties:
        title:
          type: string
        targetDate:
          type: string
        notes:
          type: string
        motivationalMessage:
          type: string
      required: [title]
    UpdateStudentGoalBody:
      type: object
      properties:
        title:
          type: string
        targetDate:
          type: string
        notes:
          type: ["string", "null"]
        motivationalMessage:
          type: ["string", "null"]
        isCompleted:
          type: boolean
...
    ExamRotation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        teacherScope:
          $ref: "#/components/schemas/ExamRotationTeacherScope"
        selectedTracks:
          type: array
          items:
            type: string
        createdById:
          type: integer
        createdAt:
          type: string
      required: [id, name, startDate, endDate, isActive, teacherScope, selectedTracks, createdById, createdAt]
    CreateExamRotationBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        teacherScope:
          $ref: "#/components/schemas/ExamRotationTeacherScope"
        selectedTracks:
          type: array
          items:
            type: string
      required: [name, startDate, endDate]
    ExamRotationTeacherScope:
      type: string
      enum: [girls, selected_tracks]
    ExamTeacherAssignment:
      type: object
      properties:
        id:
          type: integer
        rotationId:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        originalCircleId:
          type: integer
        originalCircleName:
          type: string
        originalMeetingTime:
          type: ["string", "null"]
        examCircleId:
          type: integer
        examCircleName:
          type: string
        examMeetingTime:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, rotationId, teacherId, teacherName, originalCircleId, originalCircleName, examCircleId, examCircleName, createdAt]
    ExamAssignmentInput:
      type: object
      properties:
        teacherId:
          type: integer
        originalCircleId:
          type: integer
        examCircleId:
          type: integer
      required: [teacherId, originalCircleId, examCircleId]